% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_RQ_PEC_PNEC.R
\name{calculate_RQ_PEC_PNEC}
\alias{calculate_RQ_PEC_PNEC}
\title{Calculate RQ_PEC/PNEC
Calculate Ecotoxicity Risk Quotient (RQ)}
\usage{
calculate_RQ_PEC_PNEC(input_file = NULL, AF = 1000, sheet = 1, verbose = TRUE)
}
\arguments{
\item{AF}{A vector of Assessment Factors for each contaminant (default = 1 for all).}

\item{verbose}{Logical, if TRUE, prints detailed output (default = TRUE).}

\item{PEC}{A vector of Predicted Environmental Concentrations (e.g., µg/L or mg/L).}

\item{EC50_algae}{A vector of EC50 values for algae (e.g., mg/L).}

\item{EC50_daphnids}{A vector of EC50 values for Daphnia sp. (e.g., mg/L).}

\item{EC50_fish}{A vector of EC50 values for fish (e.g., mg/L).}
}
\value{
The Risk Quotient (RQ) as a numeric value.
}
\description{
This function calculates the Risk Quotient (RQ) for contaminants based on their
Predicted Environmental Concentration (PEC), EC50 values for multiple species,
and an Assessment Factor (AF).
}
\examples{
# Example with an Excel file
# RQ <- calculate_RQ_PEC_PNEC(input_file = "path/to/your/file.xlsx", AF = 1000)

#' @export
}
